requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
python-dotenv>=1.0.0
tkinter-tooltip>=2.1.0  # For enhanced UI tooltips
customtkinter>=5.2.1    # For modern UI elements
